Show Completed Favors to confirm in HomePage
When you request a favor, someone can do this and then mark it as completed, but that person will not get the points just bt doing so, it'll be necessary that favor requester also marks it as completed (confirm completion). You will see how many favors you are missing to confirm as completed so that favor maker gets the points.